chore: display versions as ints

+++ src/PMGGenverterStatus.cpp
@@ -5,7 +5,7 @@ using namespace power_whisperpower;
 using std::endl;
 
 #define FLAG_OUT(field, flag)                                                            \
-    if (field & power_whisperpower::PMGGenverterStatus::flag) {                                              \
+    if (field & power_whisperpower::PMGGenverterStatus::flag) {                          \
         io << "  " #flag "\n";                                                           \
     }
 
@@ -34,10 +34,10 @@ std::ostream& power_whisperpower::operator<<(std::ostream& io,
     io << "PWM Scale: " << state.PWM_scale << " CNT\n";
     io << "TEST Ramp: " << state.TEST_ramp << "\n";
     io << "Inverter Model: " << state.inverter_model << "\n";
-    io << "Firmware Version: " << state.firmware_version << "\n";
-    io << "Firmware Sub Version: " << state.firmware_subversion << "\n";
-    io << "Hardware Version: " << state.hardware_version << "\n";
-    io << "Hardware Sub Version: " << state.hardware_subversion << "\n";
+    io << "Firmware Version: " << static_cast<int>(state.firmware_version) << "\n";
+    io << "Firmware Sub Version: " << static_cast<int>(state.firmware_subversion) << "\n";
+    io << "Hardware Version: " << static_cast<int>(state.hardware_version) << "\n";
+    io << "Hardware Sub Version: " << static_cast<int>(state.hardware_subversion) << "\n";
     io << "Inverter Serial Number: " << state.inverter_serial_number << "\n";
     io << "Total Work: " << state.total_work << "\n";
     io << "Maintenance Work: " << state.maintenance_work << "\n";
